Fix tsquery 500s and extract shared item filter logic
- Replace to_tsquery with websearch_to_tsquery in fuzzy item search.
  Raw user input containing tsquery operators (e.g. 'notum!') previously
  caused PostgreSQL syntax errors -> HTTP 500. websearch_to_tsquery never
  raises on user input and adds quoted-phrase/OR/-exclusion support.

- Extract duplicated filter logic from get_items/search_items into
  app/api/services/item_filter_service.py: apply_common_item_filters()
  (slot, breed, gender, faction, profession, froob, nodrop, stat_bonuses,
  stat_filters, strain), build_stat_modifier_subquery, apply_stat_filters,
  and item_detail_load_options() for the shared selectinload block.

- Endpoint-specific filters (ql range, item_class column vs stat-76
  subquery) intentionally remain in the route handlers; behavior is
  unchanged.

Verified: live smoke test against PostgreSQL 16 — old code 500s on
'notum!', new code returns 200 on all hostile inputs; both /items and
/items/search return 200 with all shared filters engaged.
